+++ src/init.h
@@ -5,8 +5,8 @@
 #define PI 3.141592
 
 
-float sample_boltzmann(float temperature);
-int amount_of_particles(float rho, float L);
-float initial_positions(float L, float* x, int n);
-int initial_velocities(float* v, int n, float temperature);
-int fill_forces_table(float *table_r, float *table_r2, float *table_f, float *table_v, float r_c, int length);
+double sample_boltzmann(double temperature);
+int amount_of_particles(double rho, double L);
+double initial_positions(double L, double* x, int n);
+int initial_velocities(double* v, int n, double temperature);
+int fill_forces_table(double *table_r, double *table_r2, double *table_f, double *table_v, double r_c, int length);
