+++ src/interaccion.h
@@ -2,8 +2,8 @@
 #include <stdlib.h>
 
 
-float min_diff(float x, float y, float L);
-float r_squared(float* x_1, float* x_2, float L);
-float potential(float r_squared, float sigma);
-float get_force_from_table(float r2, float r_c, float *table_f, float *table_r, int length);
-int update_forces(float *f, float *x, int N, float L, float r_c, float *table_f, float *table_r2, int length);
+double min_diff(double x, double y, double L);
+double r_squared(double* x_1, double* x_2, double L);
+double potential(double r_squared, double sigma);
+double get_force_from_table(double r2, double r_c, double *table_f, double *table_r, int length);
+int update_forces(double *f, double *x, int N, double L, double r_c, double *table_f, double *table_r2, int length);
